Correção Dao e Model
Correção Dao e Model.
