+++ Project-Integrator-Java/bin/.gitignore
@@ -1,4 +1,6 @@
 /controller/
 /dao/
 /model/
+/postgresql-42.2.5.jar
 /view/
+/temp/
