+++ Project-Integrator-Java/src/dao/DaoAdditional.java
@@ -1,51 +1,5 @@
 package dao;
 
-public class DaoAdditional {
-
-	int code;
-	String description;
-	float price;
-	String active;
-	int code_category;
-
-	public int getCode() {
-		return code;
-	}
-
-	public void setCode(int code) {
-		this.code = code;
-	}
-
-	public String getDescription() {
-		return description;
-	}
-
-	public void setDescription(String description) {
-		this.description = description;
-	}
-
-	public float getPrice() {
-		return price;
-	}
-
-	public void setPrice(float price) {
-		this.price = price;
-	}
-
-	public String getActive() {
-		return active;
-	}
-
-	public void setActive(String active) {
-		this.active = active;
-	}
-
-	public int getCode_category() {
-		return code_category;
-	}
-
-	public void setCode_category(int code_category) {
-		this.code_category = code_category;
-	}
+public class DaoAdditional extends DaoMaster {
 
 }
