+++ Project-Integrator-Java/src/dao/DaoCategory.java
@@ -1,24 +1,6 @@
 package dao;
 
-public class DaoCategory {
+public class DaoCategory extends DaoMaster {
 
-	int code;
-	String description;
-
-	public int getCode() {
-		return code;
-	}
-
-	public void setCode(int code) {
-		this.code = code;
-	}
-
-	public String getDescription() {
-		return description;
-	}
-
-	public void setDescription(String description) {
-		this.description = description;
-	}
-
+	
 }
