+++ Project-Integrator-Java/src/dao/DaoFilterHistory.java
@@ -1,24 +1,5 @@
 package dao;
 
-public class DaoFilterHistory {
-
-	int code;
-	int code_category;
-
-	public int getCode() {
-		return code;
-	}
-
-	public void setCode(int code) {
-		this.code = code;
-	}
-
-	public int getCode_category() {
-		return code_category;
-	}
-
-	public void setCode_category(int code_category) {
-		this.code_category = code_category;
-	}
+public class DaoFilterHistory extends DaoMaster {
 
 }
