+++ Project-Integrator-Java/src/dao/DaoGuestCheck.java
@@ -1,33 +1,36 @@
 package dao;
 
-public class DaoGuestCheck {
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import model.GuestCheck;
 
-	int code;
-	String barcode;
-	String active;
 
-	public int getCode() {
-		return code;
+public class DaoGuestCheck extends DaoMaster {
+
+	public GuestCheck findGuestCheck(int code) {
+		
+		ResultSet rs;
+		GuestCheck guestCheck = new GuestCheck();
+		
+		try {
+			openConnection();
+			
+			stmt = conn.createStatement();
+			rs = stmt.executeQuery("select code,barcode,active from guestcheck where (code=" + code + ")");
+
+			if (rs.next()) {
+				guestCheck.setCode(rs.getInt(1));
+				guestCheck.setBarcode(rs.getString(2));
+				guestCheck.setActive(rs.getString(3));
+			}
+			
+			closeConnection();
+			return guestCheck;
+
+		} catch (SQLException e) {
+			closeConnection();
+			System.out.println("ERRO SQL");
+			return guestCheck;
+		}
 	}
-
-	public void setCode(int code) {
-		this.code = code;
-	}
-
-	public String getBarcode() {
-		return barcode;
-	}
-
-	public void setBarcode(String barcode) {
-		this.barcode = barcode;
-	}
-
-	public String getActive() {
-		return active;
-	}
-
-	public void setActive(String active) {
-		this.active = active;
-	}
-
 }
