+++ Project-Integrator-Java/src/dao/DaoMaster.java
@@ -1,11 +1,11 @@
-package model;
+package dao;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-public abstract class ModelMaster {
+public abstract class DaoMaster {
 
 	Connection conn;
 	Statement stmt;
