+++ Project-Integrator-Java/src/dao/DaoRequestAdditionalItem.java
@@ -1,0 +1,5 @@
+package dao;
+
+public class DaoRequestAdditionalItem extends DaoMaster {
+
+}
