+++ Project-Integrator-Java/src/dao/DaoTableList.java
@@ -1,0 +1,5 @@
+package dao;
+
+public class DaoTableList extends DaoMaster {
+
+}
