+++ Project-Integrator-Java/src/model/AdditionalItem.java
@@ -1,0 +1,33 @@
+package model;
+
+public class AdditionalItem {
+
+	int code;
+	int code_additional;
+	int code_item;
+
+	public int getCode() {
+		return code;
+	}
+
+	public void setCode(int code) {
+		this.code = code;
+	}
+
+	public int getCode_additional() {
+		return code_additional;
+	}
+
+	public void setCode_additional(int code_additional) {
+		this.code_additional = code_additional;
+	}
+
+	public int getCode_item() {
+		return code_item;
+	}
+
+	public void setCode_item(int code_item) {
+		this.code_item = code_item;
+	}
+
+}
