+++ Project-Integrator-Java/src/model/Category.java
@@ -1,0 +1,24 @@
+package model;
+
+public class Category {
+
+	int code;
+	String description;
+
+	public int getCode() {
+		return code;
+	}
+
+	public void setCode(int code) {
+		this.code = code;
+	}
+
+	public String getDescription() {
+		return description;
+	}
+
+	public void setDescription(String description) {
+		this.description = description;
+	}
+
+}
