+++ Project-Integrator-Java/src/model/ComboItem.java
@@ -1,0 +1,42 @@
+package model;
+
+public class ComboItem {
+
+	int code;
+	int countitem;
+	int code_item;
+	int code_comboitem;
+
+	public int getCode() {
+		return code;
+	}
+
+	public void setCode(int code) {
+		this.code = code;
+	}
+
+	public int getCountitem() {
+		return countitem;
+	}
+
+	public void setCountitem(int countitem) {
+		this.countitem = countitem;
+	}
+
+	public int getCode_item() {
+		return code_item;
+	}
+
+	public void setCode_item(int code_item) {
+		this.code_item = code_item;
+	}
+
+	public int getCode_comboitem() {
+		return code_comboitem;
+	}
+
+	public void setCode_comboitem(int code_comboitem) {
+		this.code_comboitem = code_comboitem;
+	}
+
+}
