+++ Project-Integrator-Java/src/model/FilterHistory.java
@@ -1,0 +1,24 @@
+package model;
+
+public class FilterHistory {
+
+	int code;
+	int code_category;
+
+	public int getCode() {
+		return code;
+	}
+
+	public void setCode(int code) {
+		this.code = code;
+	}
+
+	public int getCode_category() {
+		return code_category;
+	}
+
+	public void setCode_category(int code_category) {
+		this.code_category = code_category;
+	}
+
+}
