+++ Project-Integrator-Java/src/model/GuestCheck.java
@@ -1,0 +1,33 @@
+package model;
+
+public class GuestCheck {
+
+	int code;
+	String barcode;
+	String active;
+
+	public int getCode() {
+		return code;
+	}
+
+	public void setCode(int code) {
+		this.code = code;
+	}
+
+	public String getBarcode() {
+		return barcode;
+	}
+
+	public void setBarcode(String barcode) {
+		this.barcode = barcode;
+	}
+
+	public String getActive() {
+		return active;
+	}
+
+	public void setActive(String active) {
+		this.active = active;
+	}
+
+}
