+++ Project-Integrator-Java/src/model/Item.java
@@ -1,0 +1,69 @@
+package model;
+
+public class Item {
+
+	int code;
+	String description;
+	float price;
+	String preparation;
+	String additional;
+	String active;
+	int code_category;
+
+	public int getCode() {
+		return code;
+	}
+
+	public void setCode(int code) {
+		this.code = code;
+	}
+
+	public String getDescription() {
+		return description;
+	}
+
+	public void setDescription(String description) {
+		this.description = description;
+	}
+
+	public float getPrice() {
+		return price;
+	}
+
+	public void setPrice(float price) {
+		this.price = price;
+	}
+
+	public String getPreparation() {
+		return preparation;
+	}
+
+	public void setPreparation(String preparation) {
+		this.preparation = preparation;
+	}
+
+	public String getAdditional() {
+		return additional;
+	}
+
+	public void setAdditional(String additional) {
+		this.additional = additional;
+	}
+
+	public String getActive() {
+		return active;
+	}
+
+	public void setActive(String active) {
+		this.active = active;
+	}
+
+	public int getCode_category() {
+		return code_category;
+	}
+
+	public void setCode_category(int code_category) {
+		this.code_category = code_category;
+	}
+
+}
