--- Project-Integrator-Java/src/model/ModelAdditional.java
@@ -1,5 +1,0 @@
-package model;
-
-public class ModelAdditional extends ModelMaster {
-
-}
