--- Project-Integrator-Java/src/model/ModelAdditionalItem.java
@@ -1,5 +1,0 @@
-package model;
-
-public class ModelAdditionalItem extends ModelMaster {
-
-}
