--- Project-Integrator-Java/src/model/ModelCategory.java
@@ -1,6 +1,0 @@
-package model;
-
-public class ModelCategory extends ModelMaster {
-
-	
-}
