--- Project-Integrator-Java/src/model/ModelComboItem.java
@@ -1,5 +1,0 @@
-package model;
-
-public class ModelComboItem extends ModelMaster {
-
-}
