--- Project-Integrator-Java/src/model/ModelFilterHistory.java
@@ -1,5 +1,0 @@
-package model;
-
-public class ModelFilterHistory extends ModelMaster {
-
-}
