--- Project-Integrator-Java/src/model/ModelItem.java
@@ -1,5 +1,0 @@
-package model;
-
-public class ModelItem extends ModelMaster {
-
-}
