--- Project-Integrator-Java/src/model/ModelRequest.java
@@ -1,5 +1,0 @@
-package model;
-
-public class ModelRequest extends ModelMaster {
-
-}
