--- Project-Integrator-Java/src/model/ModelRequestAdditionalItem.java
@@ -1,5 +1,0 @@
-package model;
-
-public class ModelRequestAdditionalItem extends ModelMaster {
-
-}
