--- Project-Integrator-Java/src/model/ModelRequestItem.java
@@ -1,5 +1,0 @@
-package model;
-
-public class ModelRequestItem extends ModelMaster {
-
-}
