--- Project-Integrator-Java/src/model/ModelTableList.java
@@ -1,5 +1,0 @@
-package model;
-
-public class ModelTableList extends ModelMaster {
-
-}
