+++ Project-Integrator-Java/src/model/TlabeList.java
@@ -1,0 +1,33 @@
+package model;
+
+public class TlabeList {
+
+	int code;
+	String description;
+	String active;
+
+	public int getCode() {
+		return code;
+	}
+
+	public void setCode(int code) {
+		this.code = code;
+	}
+
+	public String getDescription() {
+		return description;
+	}
+
+	public void setDescription(String description) {
+		this.description = description;
+	}
+
+	public String getActive() {
+		return active;
+	}
+
+	public void setActive(String active) {
+		this.active = active;
+	}
+
+}
